+++ quake3-rs/ioquake3/src/opus_1_2_1/celt/entenc.rs
@@ -392,7 +392,7 @@ pub unsafe extern "C" fn ec_enc_shrink(
         ((*_this).end_offs as libc::c_ulong)
             .wrapping_mul(::std::mem::size_of::<u8>() as libc::c_ulong)
             .wrapping_add(
-                (0 as i32 as libc::c_long
+                (0 as i32 as isize
                     * (*_this)
                         .buf
                         .offset(_size as isize)
@@ -402,7 +402,7 @@ pub unsafe extern "C" fn ec_enc_shrink(
                                 .buf
                                 .offset((*_this).storage as isize)
                                 .offset(-((*_this).end_offs as isize)),
-                        ) as libc::c_long) as libc::c_ulong,
+                        ) as isize) as libc::c_ulong,
             ),
     );
     (*_this).storage = _size;
